requirements.txt
autopep8==2.0.2
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.2
Fabric3==1.14.post1
flake8==6.0.0
mccabe==0.7.0
paramiko==2.12.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyNaCl==1.5.0
six==1.16.0
tomli==2.0.1
